requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-crispy-forms==1.14.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
phonenumbers==8.12.44
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
sqlparse==0.4.2
unicode==2.8
